+++ test-harness/tests/keeper_tests.rs
@@ -6,7 +6,7 @@ use test_harness::{
     BOB, STABLECOIN_EMODE,
 };
 
-fn supply_threshold_bps(t: &LendingTest, account_id: u64, asset_name: &str) -> i128 {
+fn supply_threshold_bps(t: &LendingTest, account_id: u64, asset_name: &str) -> u32 {
     let asset = t.resolve_asset(asset_name);
     t.env.as_contract(&t.controller_address(), || {
         let map: soroban_sdk::Map<soroban_sdk::Address, common::types::AccountPosition> = t
