+++ test-harness/tests/strategy_edge_tests.rs
@@ -51,7 +51,7 @@ fn flatten<T>(
     }
 }
 
-fn supply_position_params(t: &LendingTest, account_id: u64, asset_name: &str) -> (i128, i128) {
+fn supply_position_params(t: &LendingTest, account_id: u64, asset_name: &str) -> (u32, u32) {
     let asset = t.resolve_asset(asset_name);
     t.env.as_contract(&t.controller_address(), || {
         let map: soroban_sdk::Map<soroban_sdk::Address, common::types::AccountPosition> = t
